requirements (1).txt
# Amanat-Terra — Зависимости проекта
# Установка: pip install -r requirements.txt

# Спутниковые данные
sentinelhub>=3.9.0          # Sentinel Hub API

# Аналитика
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Визуализация
matplotlib>=3.7.0
folium>=0.14.0               # Интерактивные карты
plotly>=5.15.0               # Графики для дашборда

# Отчёты
reportlab>=4.0.0             # Генерация PDF

# Веб-интерфейс
streamlit>=1.28.0

# База данных (опционально, для production)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Геопространственные данные
shapely>=2.0.0
pyproj>=3.6.0
geopandas>=0.14.0

# Утилиты
python-dotenv>=1.0.0         # Переменные окружения (API ключи)
requests>=2.31.0
